+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/clustering/GaussianDPMM.java
@@ -16,7 +16,7 @@
 package com.datumbox.framework.core.machinelearning.clustering;
 
 import com.datumbox.framework.common.Configuration;
-import com.datumbox.framework.common.dataobjects.MatrixDataframe;
+import com.datumbox.framework.common.dataobjects.DataframeMatrix;
 import com.datumbox.framework.common.dataobjects.Record;
 import com.datumbox.framework.common.storageengines.interfaces.StorageEngine;
 import com.datumbox.framework.core.machinelearning.common.abstracts.AbstractTrainer;
@@ -154,7 +154,7 @@ public class GaussianDPMM extends AbstractDPMM<GaussianDPMM.Cluster, GaussianDPM
         /** {@inheritDoc} */
         @Override
         protected double posteriorLogPdf(Record r) {
-            RealVector x_mu = MatrixDataframe.parseRecord(r, featureIds);
+            RealVector x_mu = DataframeMatrix.parseRecord(r, featureIds);
 
             x_mu = x_mu.subtract(mean);
 
@@ -180,7 +180,7 @@ public class GaussianDPMM extends AbstractDPMM<GaussianDPMM.Cluster, GaussianDPM
         protected void add(Record r) {
             assertModifiable();
 
-            RealVector rv = MatrixDataframe.parseRecord(r, featureIds);
+            RealVector rv = DataframeMatrix.parseRecord(r, featureIds);
 
             //update cluster clusterParameters
             xi_sum = xi_sum.add(rv);
@@ -201,7 +201,7 @@ public class GaussianDPMM extends AbstractDPMM<GaussianDPMM.Cluster, GaussianDPM
 
             size--;
 
-            RealVector rv = MatrixDataframe.parseRecord(r, featureIds);
+            RealVector rv = DataframeMatrix.parseRecord(r, featureIds);
 
             //update cluster clusterParameters
             xi_sum=xi_sum.subtract(rv);